add getByYear function
